refactor baiduSearch internal util to be locally scoped
